openai/requirements.txt
# Azure OpenAI and Azure Storage dependencies
openai>=1.0.0
azure-storage-blob>=12.0.0
azure-core>=1.0.0

# Database dependencies
sqlalchemy>=2.0.0
asyncpg>=0.28.0
alembic>=1.12.0

# Async support
asyncio-mqtt>=0.13.0

# JSON and data processing
jsonschema>=4.0.0

# Environment and configuration
python-dotenv>=1.0.0

# Utilities
pathlib2>=2.3.0
typing-extensions>=4.0.0
